修复: rope_params 使用 float32 生成 complex64 (NPU 不支持 complex128 算子)
Ascend NPU 的 aclnnCat/aclnnIndex 等算子不支持 DT_COMPLEX128,
但支持 DT_COMPLEX64。源代码中 torch.float64 → torch.polar()
导致 complex128, 改为 float32 即可生成 complex64。
精度影响可忽略 — RoPE 角度计算不需要 double precision.

+++ Matrix-Game-3/wan/modules/model.py
@@ -240,10 +240,11 @@ def sinusoidal_embedding_1d(dim, position):
 @torch.amp.autocast('cuda', enabled=False)
 def rope_params(max_seq_len, dim, theta=10000):
     assert dim % 2 == 0
+    # 使用 float32 以生成 complex64 — Ascend NPU 不支持 complex128 算子
     freqs = torch.outer(
         torch.arange(max_seq_len),
         1.0 / torch.pow(theta,
-                        torch.arange(0, dim, 2).to(torch.float64).div(dim)))
+                        torch.arange(0, dim, 2).float().div(dim)))
     freqs = torch.polar(torch.ones_like(freqs), freqs)
     return freqs
